+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Oauth2 Yii2 Change Log
 ======================
 
+ * Chg: update dependencies
+ * Enh: Adding `CORS` support for `token` endpoint
  * Enh: Allow `HttpBearerAuth` and `QueryParamAuth` for method `findIdentityByAccessToken` 
 
 1.0.3 December 16, 2016